Move BasicFormatter back into class library
So it can be used by Web project by default
